Move sample project into an npm ignored folder
Fixes #124
