--- DemoWeb.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web" InitialTargets="CopyFiles">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <Target Name="CopyFiles">
-    <Exec Command="npm run build" />
-
-    <ItemGroup>
-      <!-- Because this ItemGroup is inside the target, this will enumerate
-           all files just before calling Copy. If the ItemGroup were outside
-           the target , it would enumerate the files during evaluation, before
-           the build starts, which may miss files created during the build. -->
-      <MySourceFiles Include="dist\*.*" />
-      <Copy SourceFiles="@(MySourceFiles)" DestinationFiles="wwwroot\dist\%(RecursiveDir)%(Filename)%(Extension)" />
-    </ItemGroup>
-
-    <Copy SourceFiles="@(MySourceFiles)" DestinationFiles="wwwroot\dist\%(RecursiveDir)%(Filename)%(Extension)" />
-  </Target>
-
-
-</Project>
