+++ sample/Pages/Demos/Checkboxes.cshtml
