+++ sample/Pages/Demos/Checkboxes.cshtml.cs
