+++ sample/Pages/Demos/RemovedInputs.cshtml
