+++ sample/Pages/Demos/RemovedInputs.cshtml.cs
