+++ sample/Pages/Demos/SelectInput.cshtml
