+++ sample/wwwroot/css/site.css
@@ -1,0 +1,54 @@
+html {
+  font-size: 14px;
+}
+
+@media (min-width: 768px) {
+  html {
+    font-size: 16px;
+  }
+}
+
+.btn:focus, .btn:active:focus, .btn-link.nav-link:focus, .form-control:focus, .form-check-input:focus {
+  box-shadow: 0 0 0 0.1rem white, 0 0 0 0.25rem #258cfb;
+}
+
+html {
+  position: relative;
+  min-height: 100%;
+}
+
+body {
+  margin-bottom: 60px;
+}
+
+.form-field {
+  margin-bottom: 1em;
+  border: solid 1px #eee;
+  padding: 4px;
+}
+
+.status-message {
+  font-size: 2em;
+  font-weight: bold;
+}
+
+.input-validation-error + label {
+    color: red;
+}
+
+.input-validation-valid + label {
+    color: green;
+}
+
+table.form-data {
+    margin-bottom: 10px;
+}
+
+table.form-data th {
+    text-align: left;
+}
+
+.results {
+    margin-left: 4px;
+    color: cadetblue;
+}
